Put `segment` in separate folder, and break out styles
