+++ libs/core/src/components/segmented-control/segment/segment.style.css
@@ -1,0 +1,27 @@
+:host {
+  z-index: 1;
+}
+
+button {
+  padding: 0.5rem 1rem;
+  min-width: 6.625rem;
+  text-align: center;
+  cursor: pointer;
+  border-radius: calc(infinity * 1px);
+  appearance: none;
+  background: transparent;
+  border: none;
+  cursor: pointer;
+  font-size: inherit;
+  width: 100%;
+  height: 100%;
+}
+
+button:hover {
+  background-color: color-mix(in srgb, var(--gds-sys-color-state-layers-state-black-dim1), transparent);
+}
+
+button:focus-visible {
+  outline: 2px solid #000;
+  outline-offset: 2px;
+}
