+++ libs/core/src/components/segmented-control/segment/segment.ts
@@ -1,10 +1,14 @@
 import { property } from 'lit/decorators.js'
-import { GdsElement } from '../../gds-element'
+import { GdsElement } from '../../../gds-element'
 import {
   gdsCustomElement,
   html,
-} from '../../utils/helpers/custom-element-scoping'
-import { watch } from '../../utils/decorators/watch'
+} from '../../../utils/helpers/custom-element-scoping'
+import { watch } from '../../../utils/decorators/watch'
+import { unsafeCSS } from 'lit'
+
+import { tokens } from '../../../tokens.style'
+import style from './segment.style.css'
 
 /**
  * @element gds-segment
@@ -12,17 +16,17 @@ import { watch } from '../../utils/decorators/watch'
  */
 @gdsCustomElement('gds-segment')
 export class GdsSegment extends GdsElement {
+  static styles = [...tokens, unsafeCSS(style)]
+
   @property({ type: Boolean, reflect: true })
   selected = false
 
   connectedCallback(): void {
     super.connectedCallback()
-    this.setAttribute('role', 'button')
-    this.setAttribute('tabindex', '0')
   }
 
   render() {
-    return html`<slot></slot>`
+    return html`<button><slot></slot></button>`
   }
 
   @watch('selected')
